Add files via upload
Practice with method overloading for calculation the area depending on what arguments are passed. If one value is entered, that is the assumed radius and the area is found by multiplying radius by radius by PI. 
If two values are entered, the area is found by multiplying the first value by the second.