Add files via upload
Takes 3 arguments. If any one of the three is not the same as the other two, then a message alerts the user that not all of the numbers are equal. If each argument is different, then a message alerts the user that all of the numbers are different. If the arguments are all the same, then a message alerts the user that all of the numbers are equal to each other.