Add files via upload
Determines if a cat is playing or not.
A cat is likely to play depending on the temperature and time of year.
If it is summer, a cat is more likely to play at a higher temperature than it would be during the rest of the year. (25-45 degrees)
If it is any time of the year but summer, checks if the temperature is within the range which a cat would be likely to play. (25-35 degrees)